Renamed workspace_get_blocks
Name wasn't explicit about the usage of this file
